v153 P1 引擎机制接线：元素印记/诗人旋律/磐核/faith负载档位/狂暴维持浮点
- battle_mech.py: 新增 12 个 handler（fire_mark/ice_mark/thunder_mark/element_burst
  /element_burst_all/element_burst_3/element_multi_mark/element_mark_current 元素印记体系
  + melody/melody_chant/melody_finale 诗人旋律 + guard_core_burst 磐核爆发）
- battle.py: 元素结算倍率/磐核倍率/感电连击消费（_player_skill 伤害路径）
- battle.py: faith 负载四档治疗乘区 + 每刻衰减 + 过载触发（C-18）
- battle_modes.py: maintain_cost 支持浮点（狂暴每刻 -0.6）
- core_resources.py: 狂暴维持 1→0.6；faith load_tiers/decay_per_tick；游侠 energy regen 30→18

验证：test_v59/test_v1252/test_v95_25 全绿

+++ game/core/battle_modes.py
@@ -132,7 +132,8 @@ def dual_form_tick(player: dict, logs: list | None = None) -> list:
                 out.append("🌫️ 影舞态消退，回归常态。")
         return out
     # 手动形态（狂战士狂暴/龙裔龙焰）：维护成本从资源扣（由调用方传资源，本函数只返回应扣量）
-    mc = int(d.get("maintain_cost", _cfg(_battle_cfg("dual_form"), "maintain", 1)) or 0)
+    # v153：maintain_cost 支持浮点（狂暴每刻 −0.6 层），调用方按 float 扣
+    mc = float(d.get("maintain_cost", _cfg(_battle_cfg("dual_form"), "maintain", 1)) or 0)
     if mc > 0:
         out.append({"maintain_cost": mc})
     return out
